requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
gunicorn
Werkzeug
SQLAlchemy
psycopg2-binary
python-dotenv
python-dateutil